Fixed deleting product in pending order
