--- src/utils/formatPrice.js
@@ -1,9 +1,0 @@
-export const formatPrice = (number) => {
-    const formatter = new Intl.NumberFormat('en-CA', {
-        minimumFractionDigits: 2, 
-        maximumFractionDigits: 2 
-    });
-
-    // Format the number
-    return formatter.format(number);
-}
